refactor: 레거시 제거
