--- app/src/main/java/com/techcourse/controller/manual/LogoutController.java
@@ -1,17 +1,0 @@
-package com.techcourse.controller.manual;
-
-import com.techcourse.controller.UserSession;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-import nextstep.mvc.controller.asis.Controller;
-
-public class LogoutController implements Controller {
-
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse res) throws Exception {
-        final HttpSession session = req.getSession();
-        session.removeAttribute(UserSession.SESSION_KEY);
-        return "redirect:/";
-    }
-}
