--- mvc/src/main/java/nextstep/mvc/handleradapter/ControllerHandlerAdapter.java
@@ -1,24 +1,0 @@
-package nextstep.mvc.handleradapter;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
-import nextstep.mvc.view.ModelAndView;
-import nextstep.mvc.view.View;
-
-public class ControllerHandlerAdapter implements HandlerAdapter{
-
-    @Override
-    public boolean supports(Object handler) {
-        return handler instanceof Controller;
-    }
-
-    @Override
-    public ModelAndView handle(HttpServletRequest request, HttpServletResponse response, Object handler) throws Exception {
-        String path = ((Controller) handler).execute(request, response);
-        View view = new JspView(path);
-
-        return new ModelAndView(view);
-    }
-}
